pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mega-meter-reader"
version = "0.1.0"
description = "YOLO-based meter reading fine-tuning (mega_meter_reader package)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ultralytics>=8.0.0",
    "numpy>=1.24.0,<2.0.0",
    "opencv-python-headless>=4.8.0",
    "PyYAML>=6.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
